Fix POST handling.
In a bug caught by @henryjyc, which I introduced after misreading the
documentation (and now I wonder why it works when the orchestrator isn't
dispatching requests), POST handlers expected their parameters either in
the URL or as 'form-encoded' body data, while clients expected to pass
the data in the body as a JSON object. This should now work everywhere,
though the orchestrator also needs to have related fixes applied.